refactored part of employee import
