+++ server/uploads/1673428778659test.csv
@@ -1,0 +1,8 @@
+Vorname,Nachname,Strasse,Nr,PLZ,Ort,Land,Position,Abteilung
+
+
+Byron,Gorgonzola,Cheestreet,53a,14521,Edamer,Goudaland,Controller,Controlling
+Niklas,Hardrock,Headbangerstreet,53,32141,Metalcity,RocknLand,Mechanic,Fleet
+Max,,Blastraße,10,12345,Berlin,Deutschland,Chef,
+Peter,Meister,Falschstraße,42b,54321,Berlin,deutschland,,Verkauf
+Jürgen,Schmidt,,,,,,Admin,IT
